requirements.txt
Django
django-countries
django-sass-processor
django-photologue
django-compressor
django-localflavor
django-filter
django-google-maps
django-allauth
pytidylib
psycopg2-binary
djangorestframework
-e git://github.com/weitblicker/django-rest-auth.git#egg=django-rest-auth
django-rest-auth[with_social]
Pillow>=7.0.0
libsass
num2words
markdown>=3.0
martor>=1.5.1
django-modeltranslation>=0.16.1
django-tinymce4-lite
python-slugify
django-microsoft-auth
django-haystack==3.0b1
elasticsearch>=2.0.0,<3.0.0
django-scheduler>=0.9.3
form_designer
django-el-pagination
markdown-fenced-code-tabs
rules
django_reverse_admin
django-simple-captcha
django-honeypot
# load blog posts
pandas
pymysql
sshtunnel
django-instagram
cairosvg
django-cleanup
easy-thumbnails
django-location-field